Module Cnf.Literal

A literal is an input bit to the CNF. They are constructed with the vector and bit functions, negated with (~:) and managed as a group within a Disjunction.t.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val to_string : t -> Base.string

val of_labels : Label.t Base.array -> t Base.array

val create : ?width:Base.int -> ?hidden:Base.bool -> Base.string -> t Base.array

val create1 : ?hidden:Base.bool -> Base.string -> t