Module Hardcaml_verify_kernel.Dimacs

val write_problem : Cnf.t -> Stdio.Out_channel.t -> Base.unit

Write a DIMACS formatted CNF file suitable for input into SAT solvers.

val parse_result : Base.string Base.list -> Base.int Base.list Sat.t Base.Or_error.t

Parse the result of a SAT solver. Different solves use different formats, so the parsing is quie relaxed and works with minisat, picosat and z3.

val read_result : Stdio.In_channel.t -> Base.int Base.list Sat.t Base.Or_error.t

Read the result of a SAT solver.