Module Hardcaml_verify_kernel.Label

Labels represent named variables. They contain a uid (unique identifier) which is constructed when the label is created. The name is not used for comparison.

We need to represent vectors, even though the core variable type is a single bit. As such labels also refer to their bit position. The pair (uid,bit_pos) is used for comparison in order to distinguish labels.

module Uid : sig ... end

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val to_string : t -> Base.string

include Base.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : ( t, comparator_witness ) Base__Comparator.comparator

val create : ?width:Base.int -> ?hidden:Base.bool -> Base.string -> t Base.array

val create1 : ?hidden:Base.bool -> Base.string -> t

val uid : t -> Uid.t

val bit_pos : t -> Base.int

val name : t -> Base.string

val hidden : t -> Base.bool